Lots of Translation
This commit includes a lot of the base changes needed from the Spaceworld Translation team from a few years ago to get the English characters "working". Majority of the dialogue that has been implemented is from the original translation team but I have had to implement it manually myself.

Much of the demo build has now been translated but there are a few problems that need to be fixed. Stuff like extended dialogue in English on certain maps breaks the sprites or triggers. Or the fact that it appears as though certain dialogue was never disassembled in the first place. Or that Trainer Class Names cannot be more characters in length than their JP ones for whatever reason. I also couldn't figure out how to translate the names of Pokemon or their moves without breaking a lot more of the game.

I am not sure how I am going to correct these issues yet as it seems like most of this may be beyond my skills. Going to continue to play with my options and see what happens as I go. Just wanted to commit these changes so that they exist for others to play with if they stumble across them.