Returned Route 1 Texts w/Comments
Reverting these specific changes so that the demo mode runs fully. The English code is currently in the proper places just commented below the Japanese code.

It appears that Route1P1.asm is the root of the issue as this pushes the bytes down into where Route1P2.asm is calling to for the overworld sprite data. Not sure how to fix this yet as moving these files around to different banks seems to cause more harm than good a lot of time.

Committing this so that anyone who may look at the code can see my research so far and maybe help figure out the best way to resolve this issue. Especially since the changes here may very well be ported to the full story mode patch that I am also working on.