More Translation & ROMX Bank Changes
I figured out how to move the Route 1 data to a different ROMX Bank which helped fixed the overwriting data issue. Only problem is that this is not a good solution for long term reasons.

Also, we are now running into a similar issue when working with the Player's House or Silent Hill. When translating these files, certain overworld events that are normally hidden will become shown. Currently, something is causing the brother to show in the Player's Room even though he shouldn't be seen in Demo Mode. But this specifically is not linked to any of the Player House translations. So I am not sure what is causing this exactly.

I think I need some way to figure out how to change the storage locations of these overworld events so that I can move them further down/up depending on the translation file sizes.

I have also begun changing how I work with the translation to help reduce needing to massively revert files. What I am doing now is commenting out the Japanese text and recreating the English text version just below it. This should help people working with the code be able to better understand changes that have been made without needing to reference back to old commits. This is not fully done yet and will take some time.