Converting to New Formatting & Reverting Changes
Removed the translations in text_entry.asm as they were not necessary for the demo mode. They are just comments in the code now.

Updated several more files to be in line with the new format of having both languages side by side with just one commented out.

Removing all of the translation currently (turning English code into comments below Japanese code) in an attempt to locate the bug causing a crash when going the the left in the Pokemon Summary Screen as well as the reason that the brother is showing up in PlayerHouse2F when he is not supposed to be in demo mode. Once I understand the root cause and how to fix it, I will return to trying to translate things again.