Most Changes Reverted & New Format Update Completed
I have officially swapped all of the translations to the new format.

I have also unfortunately undid almost every change from the translation in an attempt to fix the bugs that are occurring due to size mismatches. I have left the character map gfx file change, the charmap.asm changes, and included my latest moving of ROM banks in layout.link to see what exactly breaks where.

I am not sure what is causing the crash on the Pokemon Summary screen's leftmost option as it no longer is broken even without finding a direct tie in.

The brother appearing when he shouldn't in the player's room in demo mode appears that it is indeed due to the text in one of the scripts files for the PlayerHouse just extending past its bounds and interfering with the brother's spawn bounds. Will need to figure out how best to handle this issue.